refactor: move transcript extraction to module
